细节及体验调整

+++ src/main/java/org/chobit/calf/service/ScriptService.java
@@ -63,6 +63,9 @@ public class ScriptService {
 
     @CacheEvict(allEntries = true)
     public Boolean delete(int id) {
+        if (id <= 6) {
+            return true;
+        }
         return scriptMapper.delete(id);
     }
 }
